pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "second-voice"
version = "0.1.0"
description = "Recursive AI assistant for voice-driven text refinement"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "sounddevice",
    "soundfile",
    "numpy",
    "rich",
    "google-api-python-client>=2.90",
    "google-auth-oauthlib>=1.2",
    "google-auth>=2.26",
]

[project.optional-dependencies]
test = [
    "pytest",
]
gui = ["tkinter"]
dev = [
    "pytest",
    "unittest-mock",
]

[project.scripts]
second-voice = "cli.run:main"

[tool.setuptools.packages.find]
where = ["src"]
